+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2 (July 15, 2017)
+
+- Add concrete concrete tests:
+    - required flags implementation to check for the path argument/flag
+    
 ## 0.0.1 (July 15, 2017)
 
 - Add vendor packages: